Added "FadeView" to ViewContainer which sets if the container fades new views in when they are shown.
ModuleManagement view now updates the module settings section (hides or unhides) when the module is disabled or enabled.

New modules with required and optional permissions will have optional permissions checked by default.

With changes to FadeView, the CategoryTitle on a SettingsView requires that the view be built before the setting can be set.  For now, we wait until the SettingView is built before we set the CategoryTitle.

Temporarily disabled the string setting renderer in SettingView since it fails to instance as it is not thread-safe.

Tinted background and a border is now shown on SettingsView when the CategoryTitle is set.